requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-google-genai==2.0.8
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
gradio==4.44.1
rank-bm25==0.2.2
pandas==2.2.3
numpy